requirements.txt
langchain>=0.3.25
langchain-community>=0.3.25
langchain-ollama>=0.3.2
langchain-google-genai>=2.1.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
opensearch-py==2.4.2
pypdf==4.0.1
python-docx==1.1.0
docx2txt==0.8
playwright>=1.40.0
gradio>=5.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
